+++ packages/post-processors/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.0.0
+  - @volontariapp/outbox@0.7.1
+
 ## 1.0.1
 
 ### Patch Changes